Refactor ModeSelection handlers
- Add command grouping (Credentials, Mode)
- Split ModeSelectionEffectHandler into 2 focused sub-handlers
- Update factory and tests to use sub-handlers
- Replace flow{} with actions{} helper
