+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/mode/presentation/ModeSelectionEffectHandler.kt
@@ -1,92 +1,15 @@
 package space.be1ski.vibits.shared.feature.mode.presentation
 
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emptyFlow
-import kotlinx.coroutines.flow.flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
-import space.be1ski.vibits.shared.core.logging.Log
-import space.be1ski.vibits.shared.feature.auth.domain.model.Credentials
-import space.be1ski.vibits.shared.feature.auth.domain.usecase.InitializeCredentialsFromEnvUseCase
-import space.be1ski.vibits.shared.feature.auth.domain.usecase.LoadCredentialsUseCase
-import space.be1ski.vibits.shared.feature.auth.domain.usecase.SaveCredentialsUseCase
-import space.be1ski.vibits.shared.feature.memos.data.ConnectionTester
-import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
-import space.be1ski.vibits.shared.feature.mode.domain.usecase.SaveAppModeUseCase
-
-private const val TAG = "ModeEffect"
-private const val LOG_URL_MAX_LENGTH = 20
 
 class ModeSelectionEffectHandler(
-  private val connectionTester: ConnectionTester,
-  private val initializeCredentialsFromEnv: InitializeCredentialsFromEnvUseCase,
-  private val loadCredentials: LoadCredentialsUseCase,
-  private val saveCredentials: SaveCredentialsUseCase,
-  private val saveAppMode: SaveAppModeUseCase,
+  private val credentialsHandler: ModeSelectionCredentialsEffectHandler,
+  private val modeHandler: ModeSelectionModeEffectHandler,
 ) : EffectHandler<ModeSelectionEffect.Command, ModeSelectionAction> {
   override fun invoke(command: ModeSelectionEffect.Command): Flow<ModeSelectionAction> =
     when (command) {
-      is ModeSelectionEffect.Command.InitializeFromLocalConfig -> handleInitializeFromLocalConfig()
-      is ModeSelectionEffect.Command.CheckStoredCredentials -> handleCheckStoredCredentials()
-      is ModeSelectionEffect.Command.UseStoredCredentialsWithValidation -> handleUseStoredCredentials()
-      is ModeSelectionEffect.Command.ValidateCredentials -> handleValidateCredentials(command)
-      is ModeSelectionEffect.Command.SaveCredentials -> handleSaveCredentials(command)
-      is ModeSelectionEffect.Command.SaveMode -> handleSaveMode(command)
-    }
-
-  private fun handleInitializeFromLocalConfig(): Flow<ModeSelectionAction> =
-    flow {
-      initializeCredentialsFromEnv()
-      // After initialization, check if credentials were loaded
-      val credentials = loadCredentials()
-      if (credentials.baseUrl.isNotBlank() && credentials.token.isNotBlank()) {
-        Log.d(TAG, "Stored credentials found after initialization")
-        emit(ModeSelectionAction.StoredCredentialsFound)
-      } else {
-        Log.d(TAG, "No stored credentials after initialization")
-        emit(ModeSelectionAction.StoredCredentialsNotFound)
-      }
-    }
-
-  private fun handleCheckStoredCredentials(): Flow<ModeSelectionAction> =
-    flow {
-      val credentials = loadCredentials()
-      if (credentials.baseUrl.isNotBlank() && credentials.token.isNotBlank()) {
-        Log.d(TAG, "Stored credentials found")
-        emit(ModeSelectionAction.StoredCredentialsFound)
-      } else {
-        Log.d(TAG, "No stored credentials")
-        emit(ModeSelectionAction.StoredCredentialsNotFound)
-      }
-    }
-
-  private fun handleUseStoredCredentials(): Flow<ModeSelectionAction> =
-    flow {
-      // Ensure credentials are loaded from local config (in case of app reset)
-      initializeCredentialsFromEnv()
-      val credentials = loadCredentials()
-      Log.d(TAG, "Using stored credentials: baseUrl=${credentials.baseUrl.take(LOG_URL_MAX_LENGTH)}")
-      connectionTester(credentials.baseUrl, credentials.token)
-        .onSuccess { emit(ModeSelectionAction.ValidationSucceeded) }
-        .onFailure { emit(ModeSelectionAction.ValidationFailed) }
-    }
-
-  private fun handleValidateCredentials(command: ModeSelectionEffect.Command.ValidateCredentials): Flow<ModeSelectionAction> =
-    flow {
-      Log.d(TAG, "Testing connection")
-      connectionTester(command.baseUrl, command.token)
-        .onSuccess { emit(ModeSelectionAction.ValidationSucceeded) }
-        .onFailure { emit(ModeSelectionAction.ValidationFailed) }
-    }
-
-  private fun handleSaveCredentials(command: ModeSelectionEffect.Command.SaveCredentials): Flow<ModeSelectionAction> =
-    flow {
-      Log.d(TAG, "Saving credentials")
-      saveCredentials(Credentials(command.baseUrl, command.token))
-    }
-
-  private fun handleSaveMode(command: ModeSelectionEffect.Command.SaveMode): Flow<ModeSelectionAction> =
-    flow {
-      Log.i(TAG, "Saving mode: ${command.mode}")
-      saveAppMode(command.mode)
+      is ModeSelectionEffect.Command.Credentials -> credentialsHandler(command)
+      is ModeSelectionEffect.Command.Mode -> modeHandler(command)
     }
 }
